refactor: 바스켓 캐러셀 로직 훅으로 분리
