+++ apps/web/src/app/tournament/[id]/create/_components/tournament-item-basket/TournamentItemBasketCarousel.tsx
@@ -1,10 +1,10 @@
 'use client';
 
-import { useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react';
+import { useEffect, useRef } from 'react';
 
 import { toast } from 'sonner';
 
-import { Carousel, type CarouselApi, CarouselContent, CarouselItem } from '@/components/carousel';
+import { Carousel, CarouselContent, CarouselItem } from '@/components/carousel';
 import type { TournamentItemT } from '@/types/tournament';
 import { cn } from '@/utils/cn';
 
@@ -12,7 +12,7 @@ import {
   BASKET_CAROUSEL_SLIDE_SIZE_PERCENT,
   ITEMS_PER_BASKET,
 } from '../../_consts/tournamentItemBasket';
-import { getActiveBasketCount, getBasketIndexForLastItem } from '../../_utils/tournamentItemBasket';
+import { useBasketCarousel } from '../../_hooks/useBasketCarousel';
 import CarouselIndicator from './CarouselIndicator';
 import TournamentItemBasket from './TournamentItemBasket';
 
@@ -22,69 +22,27 @@ type TournamentItemBasketCarouselProps = {
   onScrolled?: () => void;
 };
 
-function TournamentItemBasketCarousel({ items = [], scrollToLast = false, onScrolled }: TournamentItemBasketCarouselProps) {
-  const [carouselApi, setCarouselApi] = useState<CarouselApi>();
-  const [currentIndex, setCurrentIndex] = useState(0);
+function TournamentItemBasketCarousel({
+  items = [],
+  scrollToLast = false,
+  onScrolled,
+}: TournamentItemBasketCarouselProps) {
+  const {
+    carouselApi,
+    setCarouselApi,
+    currentIndex,
+    activeBasketCount,
+    isCarouselEnabled,
+    handleIndicatorSelect,
+  } = useBasketCarousel({ items, scrollToLast, onScrolled });
 
-  const activeBasketCount = useMemo(() => getActiveBasketCount(items.length), [items.length]);
-
-  const prevItemCountRef = useRef(scrollToLast ? 0 : items.length);
   const prevBasketCountRef = useRef(activeBasketCount);
-
-  const isCarouselEnabled = activeBasketCount > 1;
-
   useEffect(() => {
     if (activeBasketCount > prevBasketCountRef.current) {
       toast.info('카트가 꽉 찼어요! 새 카트를 만들었어요.');
     }
     prevBasketCountRef.current = activeBasketCount;
   }, [activeBasketCount]);
-
-  /** 담기 완료 시 마지막 아이템이 있는 바구니로 이동 (링크 담기: 실시간 / 위시 담기: 페이지 재진입) */
-  useEffect(() => {
-    if (!isCarouselEnabled) {
-      prevItemCountRef.current = items.length;
-      return;
-    }
-
-    if (!carouselApi) return;
-
-    if (items.length > prevItemCountRef.current) {
-      carouselApi.scrollTo(getBasketIndexForLastItem(items.length));
-      onScrolled?.();
-    }
-
-    prevItemCountRef.current = items.length;
-  }, [carouselApi, isCarouselEnabled, items.length, onScrolled]);
-
-  /** 초기 이미지 위치 틀어짐 방지 */
-  useLayoutEffect(() => {
-    if (!carouselApi) return;
-
-    carouselApi.reInit();
-    carouselApi.scrollTo(carouselApi.selectedScrollSnap(), true);
-  }, [carouselApi, activeBasketCount]);
-
-  useEffect(() => {
-    if (!carouselApi) return;
-
-    const handleSelect = () => setCurrentIndex(carouselApi.selectedScrollSnap());
-
-    const handleReInit = () => {
-      carouselApi.scrollTo(carouselApi.selectedScrollSnap(), true);
-    };
-
-    handleSelect();
-    carouselApi.on('select', handleSelect);
-    carouselApi.on('reInit', handleReInit);
-
-    return () => {
-      carouselApi.off('select', handleSelect);
-      carouselApi.off('reInit', handleReInit);
-    };
-  }, [carouselApi]);
-
-  const handleIndicatorSelect = (index: number) => carouselApi?.scrollTo(index);
 
   if (!isCarouselEnabled) {
     return (
