+++ apps/web/src/app/tournament/[id]/create/_hooks/useBasketCarousel.ts
@@ -1,48 +1,74 @@
 'use client';
 
-import { useRef, useState } from 'react';
+import { useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react';
 
-import { BASKET_COUNT } from '../_consts/tournamentItemBasket';
+import type { CarouselApi } from '@/components/carousel';
+import type { TournamentItemT } from '@/types/tournament';
 
-const SWIPE_THRESHOLD = 50;
+import { getActiveBasketCount, getBasketIndexForLastItem } from '../_utils/tournamentItemBasket';
 
-const useBasketCarousel = () => {
+type UseBasketCarouselParams = {
+  items: TournamentItemT[];
+  scrollToLast: boolean;
+  onScrolled?: () => void;
+};
+
+export const useBasketCarousel = ({ items, scrollToLast, onScrolled }: UseBasketCarouselParams) => {
+  const [carouselApi, setCarouselApi] = useState<CarouselApi>();
   const [currentIndex, setCurrentIndex] = useState(0);
-  const touchStartX = useRef<number | null>(null);
 
-  const handleKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === 'ArrowRight') {
-      setCurrentIndex(prev => Math.min(prev + 1, BASKET_COUNT - 1));
-    } else if (e.key === 'ArrowLeft') {
-      setCurrentIndex(prev => Math.max(prev - 1, 0));
-    }
-  };
+  const activeBasketCount = useMemo(() => getActiveBasketCount(items.length), [items.length]);
+  const isCarouselEnabled = activeBasketCount > 1;
 
-  const handleTouchStart = (e: React.TouchEvent) => {
-    touchStartX.current = e.touches[0]?.clientX ?? null;
-  };
+  const prevItemCountRef = useRef(scrollToLast ? 0 : items.length);
 
-  const handleTouchEnd = (e: React.TouchEvent) => {
-    if (touchStartX.current === null) return;
-
-    const endX = e.changedTouches[0]?.clientX;
-
-    if (endX == null) {
-      touchStartX.current = null;
+  useEffect(() => {
+    if (!isCarouselEnabled) {
+      prevItemCountRef.current = items.length;
       return;
     }
 
-    const deltaX = touchStartX.current - endX;
-    touchStartX.current = null;
+    if (!carouselApi) return;
 
-    if (deltaX > SWIPE_THRESHOLD && currentIndex < BASKET_COUNT - 1) {
-      setCurrentIndex(prev => prev + 1);
-    } else if (deltaX < -SWIPE_THRESHOLD && currentIndex > 0) {
-      setCurrentIndex(prev => prev - 1);
+    if (items.length > prevItemCountRef.current) {
+      carouselApi.scrollTo(getBasketIndexForLastItem(items.length));
+      onScrolled?.();
     }
+
+    prevItemCountRef.current = items.length;
+  }, [carouselApi, isCarouselEnabled, items.length, onScrolled]);
+
+  useLayoutEffect(() => {
+    if (!carouselApi) return;
+
+    carouselApi.reInit();
+    carouselApi.scrollTo(carouselApi.selectedScrollSnap(), true);
+  }, [carouselApi, activeBasketCount]);
+
+  useEffect(() => {
+    if (!carouselApi) return;
+
+    const handleSelect = () => setCurrentIndex(carouselApi.selectedScrollSnap());
+    const handleReInit = () => carouselApi.scrollTo(carouselApi.selectedScrollSnap(), true);
+
+    handleSelect();
+    carouselApi.on('select', handleSelect);
+    carouselApi.on('reInit', handleReInit);
+
+    return () => {
+      carouselApi.off('select', handleSelect);
+      carouselApi.off('reInit', handleReInit);
+    };
+  }, [carouselApi]);
+
+  const handleIndicatorSelect = (index: number) => carouselApi?.scrollTo(index);
+
+  return {
+    carouselApi,
+    setCarouselApi,
+    currentIndex,
+    activeBasketCount,
+    isCarouselEnabled,
+    handleIndicatorSelect,
   };
-
-  return { currentIndex, setCurrentIndex, handleKeyDown, handleTouchStart, handleTouchEnd };
 };
-
-export default useBasketCarousel;
